adjust return types
