+++ Framework/NexusCpp/inc/MantidNexusCpp/NeXusFile.hpp
@@ -7,6 +7,10 @@
 #include <utility>
 #include <vector>
 
+namespace {
+static const std::string NULL_STR("NULL");
+}
+
 /**
  * \file NeXusFile.hpp Definition of the NeXus C++ API.
  * \defgroup cpp_types C++ Types
@@ -16,42 +20,7 @@
 
 namespace NeXus {
 
-/**
- * The available compression types. These are all ignored in xml files.
- * \li NONE no compression
- * \li LZW Lossless Lempel Ziv Welch compression (recommended)
- * \li RLE Run length encoding (only HDF-4)
- * \li HUF Huffmann encoding (only HDF-4)
- * \ingroup cpp_types
- */
-enum NXcompression { CHUNK = NX_CHUNK, NONE = NX_COMP_NONE, LZW = NX_COMP_LZW, RLE = NX_COMP_RLE, HUF = NX_COMP_HUF };
-
-/**
- * Type definition for a type-keyed multimap
- */
-typedef std::multimap<std::string, std::string> TypeMap;
-
-/**
- * This structure holds the type and dimensions of a primative field/array.
- */
-struct Info {
-  /** The primative type for the field. */
-  NXnumtype type;
-  /** The dimensions of the file. */
-  std::vector<int64_t> dims;
-};
-
-/** Information about an attribute. */
-struct AttrInfo {
-  /** The primative type for the attribute. */
-  NXnumtype type;
-  /** The length of the attribute. */
-  unsigned length;
-  /** The name of the attribute. */
-  std::string name;
-  /** The dimensions of the attribute. */
-  std::vector<int> dims;
-};
+static Entry const EOD_ENTRY(NULL_STR, NULL_STR);
 
 /**
  * The Object that allows access to the information in the file.
@@ -68,7 +37,7 @@ public:
   /**
    * \return A pair of the next entry available in a listing.
    */
-  std::pair<std::string, std::string> getNextEntry();
+  Entry getNextEntry();
   /**
    * \return Information about the next attribute.
    */
@@ -540,14 +509,14 @@ public:
   /**
    * Return the entries available in the current place in the file.
    */
-  std::map<std::string, std::string> getEntries();
+  Entries getEntries();
 
   /** Return the entries available in the current place in the file,
    * but avoids the map copy of getEntries().
    *
    * \param result The map that will be filled with the entries
    */
-  void getEntries(std::map<std::string, std::string> &result);
+  void getEntries(Entries &result);
 
   /**
    * \copydoc NeXus::File::getSlab(void*, const std::vector<int64_t>&,
