+++ Framework/NexusCpp/inc/MantidNexusCpp/NeXusFile_fwd.h
@@ -7,6 +7,9 @@
 
 #include "MantidNexusCpp/DllConfig.h"
 #include <iosfwd>
+#include <map>
+#include <string>
+#include <vector>
 
 typedef const char CONSTCHAR;
 
@@ -164,3 +167,50 @@ enum class NXnumtype : const int {
 
 MANTID_NEXUSCPP_DLL std::ostream &operator<<(std::ostream &stm, const NXstatus status);
 MANTID_NEXUSCPP_DLL std::ostream &operator<<(std::ostream &stm, const NXnumtype type);
+
+// forward declare
+namespace NeXus {
+
+/**
+ * The available compression types. These are all ignored in xml files.
+ * \li NONE no compression
+ * \li LZW Lossless Lempel Ziv Welch compression (recommended)
+ * \li RLE Run length encoding (only HDF-4)
+ * \li HUF Huffmann encoding (only HDF-4)
+ * \ingroup cpp_types
+ */
+enum NXcompression { CHUNK = NX_CHUNK, NONE = NX_COMP_NONE, LZW = NX_COMP_LZW, RLE = NX_COMP_RLE, HUF = NX_COMP_HUF };
+
+typedef std::pair<std::string, std::string> Entry;
+typedef std::map<std::string, std::string> Entries;
+
+/**
+ * Type definition for a type-keyed multimap
+ */
+typedef std::multimap<std::string, std::string> TypeMap;
+
+/**
+ * This structure holds the type and dimensions of a primative field/array.
+ */
+struct Info {
+  /** The primative type for the field. */
+  NXnumtype type;
+  /** The dimensions of the file. */
+  std::vector<int64_t> dims;
+};
+
+/** Information about an attribute. */
+struct AttrInfo {
+  /** The primative type for the attribute. */
+  NXnumtype type;
+  /** The length of the attribute. */
+  unsigned length;
+  /** The name of the attribute. */
+  std::string name;
+  /** The dimensions of the attribute. */
+  std::vector<int> dims;
+};
+
+/** Forward declare of NeXus::File */
+class File;
+} // namespace NeXus
